feat(landing): radar reputation chart, Try Demo Privy login, mobile menu, privacy/terms pages
- Hero: remove Proof-of-Performance status badge, swap Reputation-inputs bars for a native SVG radar chart
- Nav: floating island pill (fluid, glass), Try Demo CTA that opens Privy and routes to /onboarding, mobile hamburger menu; drop Solana chip and Launch App
- Add privacy + terms pages, FAQ JSON-LD, skip-link, LiveLeaderboard heading scale
- Waitlist: email verification + verified-referral reward tiers (backend referral_tier returns i32)
- Remove TrustStrip + brand-icons; data.ts TRUST_ITEMS dropped
- AcidButton: active press feedback

+++ server-rs/crates/api/src/routes.rs
@@ -869,7 +869,7 @@ pub async fn post_waitlist(
 
 /// Verified-referral reward tiers. Queue influence is capped in
 /// `get_waitlist_position`; fee discounts apply to live platform fees only.
-fn referral_tier(count: i64) -> (String, i64, Vec<&'static str>) {
+fn referral_tier(count: i32) -> (String, i64, Vec<&'static str>) {
     if count >= 5 {
         (
             "Arcadian III".into(), 20,
